Minor fixes

+++ src/Main.java
@@ -18,36 +18,40 @@ public class Main {
         SimpLanPlusLexer lexer = new SimpLanPlusLexer(stream);
         List<Token> lexerErrors = new ArrayList<>();
         Token token = lexer.nextToken();
+
         while (token.getType() != SimpLanPlusLexer.EOF)
         {
             if (token.getType() == SimpLanPlusLexer.ERR)
+            {
                 lexerErrors.add(token);
+            }
             token = lexer.nextToken();
         }
 
-        //System.out.println("Numero di errori lessicali: "+ lexerErrors.size());
-
-        File f = new File("out/errors.txt");
-        if(!f.exists()){
-        	f.createNewFile();
-        } else {
-        	f.delete();
-        	f.createNewFile();
-        }
-        
-        for (int i = 0; i<lexerErrors.size();i++)
+        if (lexerErrors.size() > 0)
         {
-            int errLine = lexerErrors.get(i).getLine();
-            String errStr = lexerErrors.get(i).getText();
-            int errPos = lexerErrors.get(i).getCharPositionInLine() +1;
-            int k=i+1;
-            String toWrite = "Errore "+k+": Linea "+errLine+", carattere numero "+errPos+" -> "+errStr+"\n";
-            Files.write(Paths.get("out/errors.txt"), toWrite.getBytes(), StandardOpenOption.APPEND);
+            File f = new File("out/errors.txt");
+            if(!f.exists()){
+                f.createNewFile();
+            } else {
+                f.delete();
+                f.createNewFile();
+            }
+
+            for (int i = 0; i<lexerErrors.size();i++)
+            {
+                int errLine = lexerErrors.get(i).getLine();
+                String errStr = lexerErrors.get(i).getText();
+                int errPos = lexerErrors.get(i).getCharPositionInLine() +1;
+                int k=i+1;
+                String toWrite = "Errore "+k+": Linea "+errLine+", carattere numero "+errPos+" -> "+errStr+"\n";
+                Files.write(Paths.get("out/errors.txt"), toWrite.getBytes(), StandardOpenOption.APPEND);
+            }
+                System.out.println("Errori sintattici rilevati, consultare il file /out/errors.txt");
         }
         
         
         //Exercise 2
-
